+++ NEWS.md
@@ -1,5 +1,14 @@
 # News
 
+## bizicount 1.3.0
+
+* Deprecated `scaling` parameter, as there is no reliable way to scale 
+covariates properly when transformations and interactionsare introduced to model 
+formulas.
+
+* Deprecated `na.action` as there are no methods for `bizicount` that can
+leverage alternative `na.actions` other than `na.omit`. 
+
 ## bizicount 1.2.0 
 
 ### Major